[refactor/#63] my-activites 페이지 컴포넌트 구조 개선
