+++ src/app/(header)/my-activities/page.tsx
@@ -2,7 +2,7 @@
 
 import { SideNavigationMenu } from "@/components/layout/SideNavigationMenu";
 import EmptyList from "@/assets/svgs/empty_list.svg";
-import { ActivityItem } from "@/components/ActivityHistory/ActivityItem";
+import { ActivityItem } from "@/components/my-activities/ActivityItem";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
