+++ src/components/my-activities/registration/IntroImagesPicker.tsx
@@ -1,33 +1,33 @@
 import Image from "next/image";
+import { RefObject } from "react";
 
 // Icons
 import AddIcon from "@/assets/svgs/add_icon.svg";
 import example from "@/assets/svgs/example.svg";
 import DeleteIcon from "@/assets/svgs/delete_icon.svg";
-import { RefObject } from "react";
 
-type BannerImagePickerProps = {
-  bannerImage: string | null;
-  bannerImageRef: RefObject<HTMLInputElement>;
-  handleBannerInputClick: () => void;
+type IntroImagesPickerProps = {
+  introImages: string[] | null;
+  introImageRef: RefObject<HTMLInputElement>;
+  handleIntroInputClick: () => void;
 };
 
-export function BannerImagePicker({
-  bannerImage,
-  bannerImageRef,
-  handleBannerInputClick,
-}: BannerImagePickerProps) {
+export function IntroImagesPicker({
+  introImages,
+  introImageRef,
+  handleIntroInputClick,
+}: IntroImagesPickerProps) {
   return (
     <section className="flex flex-col gap-y-[24px]">
-      <h3 className="text-2xl text-black font-bold">배너 이미지</h3>
+      <h3 className="text-2xl text-black font-bold">소개 이미지</h3>
 
-      <input type="file" ref={bannerImageRef} className="hidden" />
-      <div className="flex gap-x-[24px]">
+      <input type="file" ref={introImageRef} className="hidden" />
+      <div className="flex flex-wrap gap-x-[24px] gap-y-[24px]">
         <button
-          className="flex flex-col justify-center items-center gap-y-[30px] rounded-[12px] border border-gray-900 border-dotted w-[180px] h-[180px] cursor-pointer px-[35px] py-[35px] bg-white"
-          onClick={handleBannerInputClick}
+          className="flex flex-col justify-center items-center gap-y-[30px] rounded-[12px] border border-gray-900 border-dotted  w-[180px] h-[180px] cursor-pointer px-[35px] py-[35px] bg-white"
+          onClick={handleIntroInputClick}
           type="button"
-          aria-label="배너 이미지 등록"
+          aria-label="소개 이미지 등록"
         >
           <Image
             src={AddIcon}
@@ -41,11 +41,11 @@ export function BannerImagePicker({
           </span>
         </button>
 
-        {bannerImage && (
-          <div className="relative">
+        {introImages.map((image, idx) => (
+          <div key={idx} className="relative">
             <Image
               src={example}
-              alt="배너 이미지"
+              alt="소개 이미지"
               width={180}
               height={180}
               className="w-[180px] h-[180px] object-cover rounded-[24px]"
@@ -53,7 +53,7 @@ export function BannerImagePicker({
             <button
               className="absolute -top-3 -right-3 cursor-pointer"
               type="button"
-              aria-label="배너 이미지 삭제"
+              aria-label="소개 이미지 삭제"
             >
               <Image
                 src={DeleteIcon}
@@ -64,8 +64,11 @@ export function BannerImagePicker({
               />
             </button>
           </div>
-        )}
+        ))}
       </div>
+      <span className="text-2lg text-gray-900">
+        * 이미지는 최대 4개까지 등록 가능합니다.
+      </span>
     </section>
   );
 }
